refactor/ index.js in server folder add compression
