+++ .gitignore
@@ -1,2 +1,6 @@
 /node_modules
 /.vscode
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
