+++ client/src/API/WShelpers.js
@@ -5,7 +5,7 @@ const socket = openSocket('http://localhost:5000');
 // but will be necessary when we add async api request in index.js
 function getBotMessage(fn) {
   socket.on('bot message', msg => {
-    // console.log(msg);
+    console.log(msg);
     fn(msg)
   });
 }
