+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/chingu-coders/Voyage2-Bears-10#readme",
   "dependencies": {
+    "compression": "^1.7.1",
     "concurrently": "^3.5.0",
     "express": "^4.16.1",
     "socket.io": "^2.0.3"
