+++ server/index.js
@@ -1,0 +1,6 @@
+const http = require('http');
+const app = require('./express');
+
+const server = http.Server(app);
+
+module.exports = server;
